convert to ts
claude coded
