--- tools/eslint/daff-packages/index.js
@@ -1,5 +1,0 @@
-'use strict';
-
-module.exports.rules = {
-  'magento-driver-versions': require('./magento-driver-versions'),
-};
