+++ tools/eslint/daff-packages/index.ts
@@ -1,0 +1,5 @@
+import magentoDriverVersions from './magento-driver-versions';
+
+export const rules = {
+  'magento-driver-versions': magentoDriverVersions,
+};
