+++ tools/eslint/daff-packages/package.json
@@ -3,6 +3,11 @@
   "version": "0.0.0",
   "main": "index.js",
   "private": true,
+  "scripts": {
+    "build:copy-package": "shx cp package.json ../../../dist/eslint/daff-packages/package.json",
+    "build": "tsc -p tsconfig.json && npm run build:copy-package",
+    "postbuild": "shx rm -rf ../../../node_modules/eslint-plugin-daff-packages && shx mkdir -p ../../../node_modules/eslint-plugin-daff-packages && shx cp -r ../../../dist/eslint/daff-packages/* ../../../node_modules/eslint-plugin-daff-packages"
+  },
   "peerDependencies": {
     "eslint": "*",
     "jsonc-eslint-parser": "*"
